+++ Ya/Algebra/Definition.hs
@@ -221,3 +221,19 @@ wrapped :: forall into i ii .
 	Castable Dual into i =>
 	into i ii -> into (Supertype i) (Supertype ii)
 wrapped f = unwrap `compose` f `compose` wrap
+
+project :: forall p from into e s t .
+	Limit into U_I_I =>
+	Transformation Natural Functor from into (p (Product into) e) I =>
+	Castable Dual into (p (Product into) e s) =>
+	Castable Flat into (I t) =>
+	from s t -> into (Supertype (p (Product into) e s)) t
+project from = wrapped @into (map @Flat @from @into @(p (Product into) e) @I from)
+
+inject :: forall p from into e s t .
+	Co Limit into U_I_I =>
+	Transformation Natural Functor from into I (p (Sum into) e) =>
+	Castable Flat into (p (Sum into) e t) =>
+	Castable Dual into (I s) =>
+	from s t -> into s (Supertype (p (Sum into) e t))
+inject from = wrapped @into (map @Flat @from @into @I @(p (Sum into) e) from)